requirements.txt
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
decorator==5.2.1
distro==1.9.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
openai==1.100.2
packaging==25.0
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.5
ruff==0.12.10
scikit-learn==1.7.1
scipy==1.16.1
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
